feat: exclude views, future annotations etc. (basically the sidebar) into seperate component
